Revert "Fixed compiler warning C4267 'conversion from 'size_t' to 'type', possible loss of data'"
Seems to have quite a few warnings in 64-bit on my machine.

BUG=angleproject:1120

This reverts commit c5cf9bc47d0ee028adbbf9e9f94ca567eec601dc.

+++ src/compiler/translator/ParseContext.cpp
@@ -599,7 +599,7 @@ bool TParseContext::constructorErrorCheck(const TSourceLoc &line,
     {
         if (type->isUnsizedArray())
         {
-            type->setArraySize(static_cast<int>(function.getParamCount()));
+            type->setArraySize(function.getParamCount());
         }
         else if (static_cast<size_t>(type->getArraySize()) != function.getParamCount())
         {
